Add a generic VCS ignore support mechanism
This adds utils for the scanner to ignore files that have been marked as
ignored within the project being scanned itself.
This is a much more natural and expected outcome for the user, and
improves the accuracy of the scanner's output statistics.

This adds support for Git, Mercurial and Subversion, but more VCS can be
added in the future.

This will be plugged into the main scanner code in the next commits.