Web Client: use swh.web.client instead of our own
The web client from "upstream" has a compatible API and is capable of
the same things. So lets drop our own version and use the official one.

The swh.web.client version is expected to have a better handling of rate
limiting, and to keep improving over time.

At the time of writing of this commit, the parallel requests capability
from the dropped version is not available in
`swh.web.client.client.WebAPIClient`. However we do not use this
capability that much in practice and it is expected to be added soon in
`swh.web.client`