policy: drop everything but the swh.model one
The RandomDirSamplingPriority policy is backed by the
`swh.model.discovery` module. We standardise on it and future
improvement should happens in the `swh-model` package for the
improvements of all.

On Stefano Zacchiroli request. We preserve the code of the dropped
policy in the `legacy-policy` directory for reference. This is intended
dead code.