provenance: extend the web client with a richer method
This is probably not the final version of this as explained in the
documentation the function.

However this is good enough to start showing information in the scanner
UI and bootstrap the discussion around what the best provenance source
to selected depending of the usecases.

The query code was initially intended to provide a first prototype of
the API on WebAPIClient, however after discussing with Zack, we moved
the prototype in the scanner source itself. The intend is to have this
moving in a dedicated server side API soon anyway.