scanner: always determine known content
The known request is significantly cheaper and simpler to run.
Especially, it can be run in any order, allowing for more efficient
paraller fetching.

The origing request as we want to run request as high as possible in the
source tree, to have more accurate result.

Even when having to do fuzzier matching in the future, known which
neighbour are locally known will help having better origin answers.

So lets perform the known queries early in all cases.