provenance: issue request concurrently
Ultimately, the WebAPIClient will be able to issues request
concurrently, so let us do the same to prepare for this day. Since the
WebAPIClient rate limit logic is compatible with Threading, this should
not be an issue.

(of course, the WebAPIClient will process multiple SWHID per request
instead of the multiple request per SWHID we have currently, but that is
fine).

he new rate