Make dashboard provenance work with prefetched QualifiedSWHIDs
Previously, the dashboard was not using prefetched provenance data and
was chocking on the QualifiedSWHIDs provided.

This is now fixed and working.

We had to change the way the javascript we access data, teach json to deal
with the QualifiedSWHID, and make sure the prefetched data are used and
displayed by default.